refactor: userAgentInfo 분리
1. userAgentInfo 를 얻는 방법이 browser, node 가 다름
2. platform 에 따라 유동적으로 userAgentInfo 를 가져오게 하고 싶었음.
3. 그래서 utils  로 함수 분리
