+++ packages/utils/package.json
@@ -31,7 +31,8 @@
     "url": "https://github.com/boostcamp-2020/Project11-A-Web-FE-Performance-Monitoring-SDK/issues"
   },
   "dependencies": {
-    "@santry/types": "^1.1.0"
+    "@santry/types": "^1.1.0",
+    "ua-parser-js": "^0.7.22"
   },
   "gitHead": "aab5bb35a6315ce644cc97e33b2a7cc34fc29e12"
 }
