+++ packages/utils/src/addUserAgentInfo/index.ts
@@ -1,0 +1,1 @@
+export { addUserAgentInfo as browserUserAgentInfo } from './browser';
