+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 export { getGlobalObject } from './getGlobalObject';
 export { parseDsn } from './parseDsn';
+export { browserUserAgentInfo } from "./addUserAgentInfo"
